+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Do not show Register button until we are sure that registration is available for the event [#3671](https://github.com/rokwire/illinois-app/issues/3671).
 
 ## [5.0.66] - 2023-09-13
 ### Fixed